pyproject.toml
[tool.poetry]
name = "completos_bot"
version = "0.1.0"
description = ""
authors = ["Alejandro Pimentel <apimentel@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.15.0"
python-telegram-bot = "20.0a1"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
